Question3/Story_generator/requirements.txt
streamlit==1.14.0
google-generativeai==0.3.1
python-dotenv==0.19.1
